+++ YLMoment/YLMomentObject.h
@@ -253,6 +253,58 @@ static NSString * const YLMomentIso8601Format = @"yyyy-MM-dd'T'HH:mm:ssZ";
  */
 - (BOOL)isEqualToMoment:(YLMoment *)anotherMoment;
 
+/**
+ * @abstract Returns a Boolean value that indicates whether a given object
+ * is an `YLMoment` object and exactly equal the receiver.
+ * @param anotherMoment The moment to compare with the receiver.
+ * @return YES if the anotherMoment is an `YLMoment` object and is exactly
+ * equal to the receiver, otherwise NO.
+ * @since 0.9.0
+ */
+- (BOOL)isSameMoment:(YLMoment *)anotherMoment;
+
+/**
+ * @abstract Returns a Boolean value that indicates whether a given object
+ * is an `YLMoment` object and exactly equal the receiver for the given
+ * calendar unit equal or larger.
+ * @param anotherMoment The moment to compare with the receiver.
+ * @param calendarUnit The calendar unit to compare against
+ * @return YES if the anotherMoment is an `YLMoment` object and is exactly
+ * equal to the receiver for the given calendar unit equal or larger,
+ * otherwise NO.
+ * @since 0.9.0
+ */
+- (BOOL)isSameMoment:(YLMoment *)anotherMoment forUnit:(NSCalendarUnit) calendarUnit;
+
+/**
+ * @abstract Returns a Boolean value that indicates whether a given object 
+ * is an `YLMoment` and is of a later date than the receiver's.
+ * @param anotherMoment The moment to be compared with the receiver.
+ * @return YES if the anotherMoment is an `YLMoment` object and of a later 
+ * date than the receiver, otherwise NO.
+ * @since 0.8.0
+ */
+- (BOOL)isBeforeMoment:(YLMoment *)anotherMoment;
+
+/**
+ * @abstract Returns a Boolean value that indicates whether a given object
+ * is an `YLMoment` and is of an earlier date than the receiver's.
+ * @param anotherMoment The moment to be compared with the receiver.
+ * @return YES if the anotherMoment is an `YLMoment` object and of an earlier
+ * date than the receiver, otherwise NO.
+ * @since 0.8.0
+ */
+- (BOOL)isAfterMoment:(YLMoment *)anotherMoment;
+
+/**
+ * @abstract Returns `BOOL` when comparing if a given moment's date is between two other moments.
+ * @param startMoment The starting reference moment being compared against.
+ * @param endMoment The ending reference moment being compared against.
+ * @return YES if the anotherMoment is an `YLMoment` object and of an later
+ * @since 0.8.0
+ */
+- (BOOL)isBetweenMoments:(YLMoment *)startMoment andEndMoment:(YLMoment *)endMoment;
+
 #pragma mark - Proxy Method
 /** @name Proxy Method */